Extracted CurveParameters for better readability.
